refactor(nvim): move some Vimscript out of a Lua string
Because of the weird way undotree chooses to accommodate overrides, we
still need some Vimscript, to overwrite a global function, but from
there we can call into Lua.

Note that sticking the mappings in an "undotree" `ftplugin` would be the
"right" way to do this, but that doesn't work because undotree sets the
filetype and then overwrites the mappings you set up, then calls the
global function.