feat(dotfiles): set up binding to show clipboard contents
At times I am paranoid* that I am about to paste the wrong thing in a
public channel. For example I might have tried to grab something out of
a text document, but it silently fails for whatever reason (maybe I
pressed the wrong key), and so I end up pasting the thing that I
previously put in the clipboard, which could be my bank account number,
a confession, or anything.

So, set up Alt (that is Option) + V to preview the clipboard contents.
As far as I can tell, Alt+V's default behavior is not that useful; in my
testing it seems to just insert "√".

[*]: But, if they really _are_ out to get you, is that paranoia?