refactor(pacman): bury systemd config files deeper in "files/"
So that I don't have a mix of "top-level files that going to get shoved
into an /etc subdirectory" with "nested files whose paths reflect where
they are going to go under /etc".

Overall, looks like this now:

    files
    └── etc
        ├── pacman.conf
        └── systemd
            └── system
                ├── resume@.service
                └── suspend@.service

Note that "resume" is still unused, as explained in 59290291f45f3c33. I
haven't noticed my keyboard layout getting unset in a while now, and I
think it's because the keyboard is no longer going through the monitor's
incredibly flakey built-in "KVM" USB hub (I'm using an external hub
now). I'll remove the file in the next commit seeing as it will
apparently never be used; I am just including it in this move commit so
that its final resting place in the Git history is next to its cousin.