docs(dotfiles): note how to set hostname on macOS
Not sure if related, but after a kernel panic (and an additional
reboot), my work Mac is coming up without a hostname (ie. `hostname`
prints an IP address, and `hostname -s` is even less helpful than that,
printing "192" 🤦). In the Sharing pref pane the name is there as it
always has been.

Seems mostly people ask about this in the context of public WiFi
networks, but the answer still seems valid:

    https://apple.stackexchange.com/a/127405/158927