refactor(dotfiles): [demo] teach `clip` to use OSC 52 escape sequence
Both Kitty and iTerm2 grok the OCS 52 escape sequence out of the box, so
we can use that instead of Clipper in those terminals. We keep the
Clipper-based code path around as a fallback (for example, would be used
in Terminal.app or other terminals).

Labeling this as a demo though because it probably needs more work. It
works fine with small copies, but not big ones (eg. `cat
/usr/share/dict/words | clip`). Probably need to break it up into chunks
for that to work. I'm committing this so that it can be found in the
history, but I'll revert it in the next commit; will revisit it when I
have time to do a deeper investigation.

Via:

- https://github.com/wincent/clipper/issues/26#issuecomment-898799659
- https://www.reddit.com/r/vim/comments/k1ydpn/a_guide_on_how_to_copy_text_from_anywhere/
- https://github.com/roy2220/osc52pty