chore: revert "refactor(dotfiles): [demo] teach `clip` to use OSC 52 escape sequence"
As noted in the previous commit, this works, but not for large inputs
(tested in both Kitty and iTerm2). Even `cat /usr/share/dict/words |
kitty +kitten clipboard` doesn't behave well (although I am on a
several-weeks-old Kitty build on this machine, so it is possibly it
works better on current `master`; I'll try that out later), which
inclines me to think this would be a bit of a rabbit hole to go down.

This reverts commit bf75f40bdc2f2f9a0343ac422e4a61c1542e96c6.