refactor: don't use AppleScript in karabiner-kill flow
As described in:

    https://wincent.com/snippets/1170

just got locked out at the lock screen with SpaceFN jammed on and
couldn't kill Karabiner-Elements because of a permissions dialog that I
couldn't see.

This is the same dialog described here:

    https://github.com/pqrs-org/Karabiner-Elements/issues/1573#issuecomment-425630564

Now, I can't remember whether it was specifically
`karabiner_console_user_server` that was mentioned in the dialog I saw,
but in reading the rest of the thread I saw I didn't need to use
AppleScript to run the script in the terminal, and could use `open` and
a `.command` extension instead.

So, that might not prevent the permissions dialog from showing up in the
future, but it still seems like a worthy refactoring to do.