feat(contrib): add /boot/{a,b}.sh scripts on Arch
Following up on the idea mentioned in the previous commit, we now have
"a.sh" and "b.sh" scripts that can be used to set up EFI boot entries
again.

So, after a BIOS update you can either go through the procedure
mentioned at:

    https://wincent.com/wiki/Arch_Linux_cheatsheet

ie. typing some long `efibootmgr` incantations.

Or, you can:

1. Boot off the installation media into a UEFI shell.
2. Boot off the hard drive with `FS1:` followed by `a`.
3. Run `sudo bash /boot/b.sh` (and `sudo bash /boot/a.sh`) to put the
   entries back the way they were before.

Of course, I probably have the names back to front. I mapped "a" to
"linux" because that's the preferred one, and "b" to "linux-lts" because
that's only a fallback, but the way EFI entries are created, you
actually have to run "b before "a", so maybe that's a sign I should just
combine them into a single script. Whatever; I don't do this that often
so it's not probably something I have to agonize over.