feat(nvim): set up some projections for Rails-like projects
Not sure whether I want to go with something very minimal like this, or
drag in all of the vim-rails plug-in, which is quite large. I _think_
I'm going to do the latter, but I want at least a simple example in the
Git history to go back to in case I end up changing my mind later on.