refactor(nvim): use vim-rails
As mentioned in the previous commit, not sure if I want to go the
heavy-weight plug-in route or not, but am at least going to give it a
try. I did used to use vim-rails waaaaay back; it was vendored in the
very first commit of this repo (61a7e2a830ed, ".vim/.vimrc: initial
import", May 2009), until it finally got removed in in e9a1d032a9f68
("vim: remove vim-{haml,rails}", October 2015).

* aspects/nvim/files/.config/nvim/pack/bundle/opt/vim-rails 0000000...03a5c3e (1612):
  > Ensure we're back in magic mode after embedding `define_pattern()`