chore: bump node version from 14.17.0 to 16.8.0
Via the releases page:

    https://nodejs.org/en/about/releases/

v14 is the "Active LTS" release, and came out on 2020-04-21. It will hit
end-of-life on 2023-04-30, so no hurry to get off it.

v16 is the "Current" release, and came out on 2021-04-20. It will hit
end-of-life on 2024-04-30.

So yeah, no hurry to move, but I may as well update about once a year
and stay on whatever is "Current". (This partly prompted because I just
opened a PR to a project that wants me to use Node v16, and if I can
just have that version hanging around on my system already it makes life
easier for me).

Test plan: `./install` and see that things still seem to be working.