feat(dotfiles): add `gg go` shortcut
Based on the same idea as the old `portool go`:

    https://github.com/wincent/wincent/blob/e4aca1cf904d7e279c893d3df0b2d6e062897e95/aspects/dotfiles/files/.zsh/liferay/bin/portool-helpers/go

This one is in some sense simpler because I don't have as many things I
want to run, but it does have to deal with Darwin and non-Darwin (ie.
Codespaces) environments, which have repos checked out in different
places, and different capabilities/roles.