fix(nvim): unbreak <Leader>r mapping
I must have broken this months ago without noticing. The gotcha here is
that we need to support both `<expr>` and non-`<expr>` mappings.
Examples of the former category include some autocompletion-related
ones. For those, we evaluate the (VimScript) expressions, which happen
to be calls into Lua (eg. `v:lua.blah()`) and the return value ends up
being the RHS of the mapping.

The second category are ones like the `<Leader>r` mapping. For those we
have to get into a command context and call a Lua function (ie. `:lua
do_something()<CR>`).