fix(nvim): make ":OpenOnGitHub" work in Codespaces again
The image I'm working on very recently added a "/usr/bin/open" symlink,
pointing to "/etc/alternatives/open", which in turn points at
"/usr/bin/xdg-open". So that means that when `ssh`-ed into a Codespace,
`executable('open')` would evaluate to `1`, and we'd futilely try to
open the link.

Now, in the long run, I probably want to do something cleverer here, so
that I can use `open` on both macOS and my Linux box, but skip it in a
Codespaces context, but for now I am just unbreaking this as
straightforwardly as possible: instead of checking for the existence of
an executable, we get its path, resolve symlinks, and grab the final
path component; if it's "open" then we use it, but if it is anything
else (eg. "xdg-open"), we don't. Note that if the executable is missing,
`exepath` will return an empty string, `resolve` will return an empty
string too, and `fnamemodify` will return -- you guessed it -- an empty
string as well.