refactor(dotfiles): fold "git zap" and "git zip" into one script, "git sync"
As noted previously, I always run these two together, so it makes sense
for them to be melded into one script.

Fun fact: before I started this, I thought I was going to need an
associative array, so I switched over to Zsh. In the end, I didn't need
it, but I left it in here to: (a) show my work, and (b) have an example
of how to do some useful things in Zsh, like find out the path of the
currently script (like you can in Bash with `$BASH_SOURCE`).

In the next commit, I'll delete the unneeded/unused associative array,
but I'm keeping the script in Zsh for now.