fix(nvim): enable tmux plugin in zen-mode.nvim
I had this back to front because I misread the README:

    https://github.com/folke/zen-mode.nvim

which says:

    tmux = { enabled = false }, -- disables the tmux statusline

ie. I thought `enabled = false` meant "make it look like tmux isn't
enabled".

But it doesn't work. Inspection of the source code:

    https://github.com/folke/zen-mode.nvim/blob/9ae43df6ee4d6c3d47aea057086c9e56fca58234/lua/zen-mode/view.lua#L21

reveals that you need `enabled = true` to turn on the tmux plugin inside
zen-mode, and then it will do its thing:

    https://github.com/folke/zen-mode.nvim/blob/9ae43df6ee4d6c3d47aea057086c9e56fca58234/lua/zen-mode/plugins.lua#L66-L67

and turn off the status line and zoom the current pane:

    vim.fn.system([[tmux set status off]])
    vim.fn.system([[tmux list-panes -F '\#F' | grep -q Z || tmux resize-pane -Z]])

which is exactly what I wanted it to do.