feat(nvim): add binding for doing file-wise movement through jump list
This is really just a draft which I am committing in order to document
it; it seems that the feedkeys call works, but it's side-effects
don't seem to take effect immediately. That means that when we call
`nvim_win_get_cursor` and `nvim_buf_get_name` we get the old results
back.

In the next commit I am going to experiment with adding a small delay
(via `defer_fn`) to see if that allows the code to "see" the
side-effects.