fix(nvim): make jump list step function to overcome feedkeys latency
As mentioned in the parent commit, we have to wait a bit to actually see
the effects of the feedkeys call.

But as noted in the "BUG:" comment, this is still edge-casey enough that
I probably have to resort to parsing the `:jumps` output in order to
reliably detect forwards progress (and if I am going to do that, maybe I
can dispense with reading the buffer name entirely and just enqueue a
number of feedkeys calls directly; maybe even a single call containing
repeated presses).