fix(nvim): make file-wise jumping use getjumplist()
Addresses the problems noted in the previous commit. Namely, the "BUG:"
comment that points out that invalid line numbers can break all the
things. This is nicer anyway, as we're doing all the jumps via a single
mapping, and we should even be able to jump through multiple distinct
unnamed buffers now (not that those come up particularly frequently, but
anyway).