feat(fig): make most operations return an `OperationResult`
We can make this a richer return type in the future if it proves
useful, but for now we return a simple string like "ok", "changed",
or "skipped". The only operation that don't do this, for now, is
`command()` because that already returns something more "interesting"
for its return value.

Closes: https://github.com/wincent/wincent/issues/105