chore(nvim): remove stale lua-language-server instructions
Seeing as lua-language-server regularly pegs the CPU for me, and it's
been a while since I built it, going to see whether a new version fixes
it. This change from November 2020:

    https://github.com/sumneko/lua-language-server/issues/247

looks promising, and I think my build predates that... Fingers crossed.

Built using:

    cd $path_to_clone_of_lua_language_server_repo
    git pull
    git submodule update --recursive
    cd 3rd/luamake
    ./compile/install.sh
    cd ../..
    ./3rd/luamake/luamake rebuild

Then back out unsolicited changes to `~/.zshrc` 🤦 and update path in my
config to the built executable (the binary in the `macOS/` subdirectory
is there, but won't boot, failing to load "bee.so" which is right next
to it).