fix(dotfiles): make "emp something something" not blow up
The search facility here isn't very good (we can only search by login,
not name), but the least we can do is not blow up if somebody does try
to search for "foo bar".

So, convert spaces to pluses before trying to open the URL. This
prevents the `open` tool from complaining with:

    The file $absolute-path-to-current-directory/$some-url?login=foo bar does not exist.

ie. note how the space in the "foo bar" causes the tool to treat the
argument as a non-URL, and it ties to open that bizarre mongrel
formed by mashing the absolute path to the current directory with the
URL that we're trying to open.