refactor(dotfiles): override Git's merge.conflictStyle differently
Because the way I was trying to do it in 5de84b453699c976f7e2a91e262116
doesn't work. Reported the problem on the Git mailing list here:

    http://public-inbox.org/git/ee1dd453-e698-440a-911b-d14389e33715@beta.fastmail.com/

But even if it does get fixed, the fix will only apply in new releases,
so we have to work around it here.

The problem is that, when running an old version of Git, `git
checkout` dies immediately on seeing an unrecognized value for
`merge.conflictStyle` in `~/.gitconfig`, even though we later override
it with a backwards-compatible value in `~/.gitconfig.d/codespaces`, and
even though `git config merge.conflictStyle` reports the overridden
value.