feat(nvim): try to use current branch in `:OpenOnGitHub`
Will see if this is annoying or not. Because if you're on "main" or
"master" it is going to be nice, but if you are on
"topic-branch-that-maybe-i-have-not-pushed-yet", it could become
irritating. If it does, I'll look at making this even smarter, like
trying to detect the base branch if the remote doesn't seem to have a
branch matching the topic.