chore: rename user account from "glh" to "wincent"
Parent commit made me think about this and I figure, why not? "wincent"
is the username I used historically from the first days of Mac OS X
(including prior to the first public beta in 2000) on all of my personal
machines. From 2011 to 2013 I worked for `$startup` and had my first
employer-provided machines, and IIRC the username I used on those was
"wincent".

Then at `$megaCorp`, "glh" was my username from 2014 to 2018, and I
ended up using that on my personal laptop, which I acquired in 2017, for
consistency. When I got a Linux desktop in 2020 I again opted for "glh"
for consistency.

But there was a fly in the ointment: at `$nextJob` they had given
me "greghurrell" as a username (from 2019 to 2020), and finally at
`$currentJob` (2021 to current) it's again "wincent".

All of which is to say, it seems I can and may as well just go back to
using "wincent" everywhere again: personal laptop, Linux desktop, and
work laptop.