refactor(dotfiles): tweak i3-gaps to fit screenkey
Turning off `inverse_outer` because it messes with my plan to make
screenkey not overlap with other windows.

    i3-msg gaps bottom all set 70

makes sure that the windows don't overlap, but when `smart_gaps
inverse_outer` is in effect, it only works on workspaces with exactly
one window on them. As I had gaps set to 0 anyway, there is no point in
having this setting, so I am just commenting it out.

Added a `gaps` function so that I don't have to remember that `i3-msg`
incantation.