fix(dotfiles): stop leaking HOSTNAME variable
This variable could be useful, but I don't want to accidentally set it
globally in a file like this one.

It was added in 9d3e4a539363d548e ("chore: make some work-specific
config available on personal machine", Sep 17, 2019) and was already
leaking globally then. The last reference to it was removed in
6465fa42bf6e5728b ("chore: remove Liferay-specific config and tools",
May 6, 2021).

So, remove it, and use an anonymous local function so that any variables
I add here in the future don't wind up leaking, provided I just follow
the pattern.