feat(systemd): set up Pulseaudio null sink automatically via systemd
According to the internet, there are a million ways to do this, but
because I am currently using Pulseaudio compatibility mode in PipeWire,
this looks like the easiest (the alternatives all require explicit
PipeWire configuration, which would mean that I would have to stop
pretending that PipeWire doesn't exist).

Note that I tweaked the way handlers get run because I wanted to notify
multiple handlers as a result of a single task running. So, you can now
notify an array of handlers. Additionally, I want handlers to run in
notification order, not registration order. Added "tests" (ie. code in
the "meta" aspect) to show this working:

- We register three handlers in this order: "first", "second", "common".
- We then run two tasks that notify ["common", "first"] and ["common",
  "second"] respectively.
- Under the old system, we would run "first", "second", "common",
  because that is the registration order.
- Under the new system, we run "common", "first", "second", because that
  is the notification order.

The old implementation was the way it was because that seemed like the
simplest thing to build, but as soon as I added the ability to notify an
multiple handlers using array syntax, it became clear that it is too
easy to wind up with counter-intuitive results (you'd definitely expect
`notify: ['a', 'b']` to run `'a'` then `'b'`), hence this refactor.

Here's an excerpt from the "meta" run showing this working:

    [notice]  Task: meta | don't notify a handler
    [info]    Skipped: command `mkdir /etc`
    [notice]  Task: meta | notify a handler
    [notice]  Changed: command `true`
    [notice]  Task: meta | notify multiple handlers...
    [notice]  Changed: command `true`
    [notice]  Task: meta | ...across multiple tasks
    [notice]  Changed: command `true`
    [notice]  Task: meta | set repo permissions
    [info]    Ok: file /home/wincent/code/wincent
    [notice]  Handler: meta | handle command
    [notice]  Changed: command `true`
    [notice]  Handler: meta | handle common
    [notice]  Changed: command `true`
    [notice]  Handler: meta | handle first
    [notice]  Changed: command `true`
    [notice]  Handler: meta | handle second
    [notice]  Changed: command `true`
    [notice]  Handler: meta | should not to be called
    [info]    Skipped: handler meta | should not to be called