perf(nvim): mostly fix laggy screen updates when indenting/dedenting
For some reason, dedenting is slow even without indent-blankline
running. The remapping of "<" in normal mode makes it snappy though.
VISUAL mode remains a dog (but like I said, it's a dog even without
indent-blankline). It seems to be a clash with a mapping with an
ambiguous prefix (<Nop>), so I'll reconfigure that in the next commit.