perf(nvim): avoid delays caused by use of <Nop> on LHS of mappings
It's only legit on RHS, where you can use it to map something to do
nothing. But if you want to stop something being mapped, you can't use
it on the LHS; it ends up being parsed as "<" + "n" + "o" + "p" + ">",
and interfering with other mappings that have a "<" prefix.

"Solution": Bind to another key that I can't don't type instead.