chore(nvim): tweak tcomment mappings
I have to give a session on Vim in which I want to show some useful
built-in mappings, like `gu{motion}`. So, get rid of my custom `gu` for
TComment, in favor of `gC`, to avoid the clash.

ie. with this, we now have `gc` to toggle comments, and `gC` to
explicitly turn them off.

While I'm at it, removed `tcomment_textobject_inlinecomment` setting
which is no longer mentioned in the docs, so I suspect that it does
nothing.