fix(dotfiles): work around shell initialization race condition
This has been driving me crazy for a while now, that I would see
sometimes, but not always, this when I opened a shell:

    ln: colors.conf: File exists

despite the `-f` in `ln -sf`.

I believe the cause is that `ln -sf` first unlinks and then links in the
case of an existing entry (ie. it cannot do an atomic "overwrite"
operation). So, when you launch two shells at a time, something which
can happen pretty easily when a `.tmux` script runs and sets up multiple
splits, there is a chance that two shells will race to perform the
following steps (one possible ordering arbitrarily selected out of many
and shown here):

1. Shell "A" checks to see if "colors.conf" exists; it does.
2. Shell "B" checks to see if "colors.conf" exists; it does.
3. Shell "B" unlinks "colors.conf" and succeeds.
4. Shell "A" unlinks "colors.conf" and fails (silently).
5. Shell "A" creates new "colors.conf".
6. Shell "B" tries to create new "colors.conf" and fails (noisily).