refactor: add `arch`/`darwin`/`debian` helpers
As explained in the related issue, I don't know if these things belong
in the Fig core itself, so I'm adding a mechanism here for adding
arbitrary helper functions to a top-level "helpers.ts" file.

The three helpers I'm adding here allow me to DRY up the common pattern
of running specific tasks only on a specific platform. It works well for
this limited case; if we ever want an expressive way of saying "run only
on personal Darwin machines but not work Darwin machines", then we'll
need to look for something else, but at the moment there is no pressing
need.

Closes: https://github.com/wincent/wincent/issues/125