feat(fig): add support for SUDO_ASKPASS
We need to be able to do passwordless sudo on codespaces now, so I
figure this is the best way to provide a means for doing so (we'll just
set `SUDO_ASKPASS` to a program that emits an empty string).

Without this, we'll fail the first time we need to do sudo during the
installation, because we'll call `prompt()` (to get the password), but
`NON_INTERACTIVE` will be set. In my test run, this manifested in the
logs as a failure to `apt-get install` something:

    [error]   Failed: command `apt-get install -y bfs` failed

Alternatives considered:

- Ability to pass in password via environment variable or command line
  option (rejected because it encourages insecure practices, even if
  _we're_ only using it for the benign case of passing an empty string).

- Ability to indicate via an environment variable or command line switch
  that we don't want to provide a password (rejected because it seems
  very one-off and hackish, lacking in flexibility compared to the
  chosen solution).

In my local testing I checked to see this working with absolute paths
like `/bin/ls` and names only (ie. `ls`). Not sure whether `spawn` is
using the `PATH` under the hood (I'd hope so) or a hard-coded list, but
we'll see when we try this out for real on a codespace, I guess.