feat: add "node" aspect to codespaces
Previously I have only flirted with `typescript-language-server` on
codespaces, not wanted to conflict with the already-installed node on
there. My main concern is that some webpack build process (or similar)
might end up picking up on the newer (v18.0.0) version installed at:

    /home/vscode/n/bin/node

and run into some divergent breaking behavior relative to the older
version at:

    /usr/bin/node

But there's really only one way to be sure and that's to try it, so here
we go.