refactor(nvim): remove commented-out LSP config
I had this in there so I wouldn't forget how to emit debug output to a
log file, but I haven't needed to do that in recent memory, so now I
consider this to be just clutter.

The default config:

    https://github.com/neovim/nvim-lspconfig/blame/21102d5e3b6ffc6929d60418581ac1a29ee9eddd/lua/lspconfig/server_configurations/tsserver.lua

(obviously) already works without this config, already using both
`typescript-language-server` and `--stdio`, so let's just get rid of it.