refactor(nvim): switch to new keymap API provided by Neovim v0.7.0
Was added in:

    https://github.com/neovim/neovim/commit/6d41f65aa45f10a93ad476db01413abaac21f27d

The API is extremely similar to what I had, so the changes are very
small:

- Instead of mode-specific functions that correspond to `nmap`,
  `nnoremap` etc, the new API takes an explicit `mode` (really, `modes`)
  argument as the first parameter. So:

      wincent.vim.nmap(lhs, rhs, opts)

  becomes:

      vim.keymap.set('n', lhs, rhs, opts)

- Additionally, everything is `noremap` by default, unless its a
  `<Plug>` mapping, so we can basically forget about the distinction
  between `nmap` and `nnoremap` (etc) because it will do the right thing
  for us.

Finally, note that there are a couple of places where I wrapped a
function reference:

    wincent.mappings.leader.cycle_numbering

with an anonymous forwarding function:

    function() wincent.mappings.leader.cycle_numbering() end

to stop us from eagerly loading the corresponding file before the
mapping is actually used.

Lots of changes here, but all fairly simple, so hopefully I didn't break
anything inadvertently along the way.

Closes: https://github.com/wincent/wincent/issues/126