feat(nvim): [demo] use choice nodes in "imp" snippet
I don't actually like the ergonomics of this, so I'm going to back it
out in the next commit, but I wanted to at least show how it is done.

`<C-u>` to select from choices and thus produce "import" or "import
type". In practice, I think it feels better to just have an "impt"
snippet to expand to "import type". Failing that, I can put some smarts
in here so that if you type "type {Something}" in the `ModuleName`
position, it will do the right thing and strip off the "type {...}" over
on the RHS.