fix(codespaces): adapt to new hostname format
I've noticed my commits getting created with the wrong email address
because the hostnames used to be of the form:

    codespaces_${HEX_STUFF}

and now they are of the form:

    codespaces-${HEX_STUFF}

Now, I could make this even laxer, and perhaps I should, but at least
to start with I am doing the minimal thing.

Without this change, my `hostHandle` ended up being
`codespaces-${HEX_STUFF}` instead of the desired `codespaces`. As such,
my `~/.gitconfig` would up with this in it:

    [include]
            path = .gitconfig.local
            path = .gitconfig.d/codespaces-${HEX_STUFF}

instead of the desired:

    [include]
            path = .gitconfig.local
            path = .gitconfig.d/codespaces

As such, the necessary include file wasn't being loaded.