feat(nvim): hack together handle-based completion for commit messages
I'll doubtless end up tweaking this, but for now, only in commit
messages, you can type "@" to autocomplete a GitHub handle. The idea is
you do:

    cob<tab>

and it will expand to (with `@handle` selected):

    Co-Authored-By: @handle
                    ^^^^^^^
                       1

You can type "@" or "@wincent" or "@hurrell" and it will offer
completions like:

    Co-Authored-By: Greg Hurrell <wincent@github.com>

This isn't perfect (ie. you can't type "@greg hurr" and tab complete
from there; as soon as you hit space you lose the autocompletion menu)
but is is smart enough to not pop up the menu if you are typing a @ in
the middle of an email address (eg. "user@example.com" will not trigger
the pop-up). It will only trigger if "@" is preceded by space, or at the
start of a line (and to be honest, we probably don't even need the
latter although it doesn't hurt).

Depends on "~/.github-handles.json" existing on disk. Not committing
that to the repo because I don't want to leak folks' handles and names.
Format is like this:

    {
      "wincent": "Greg Hurrell <wincent@github.com>",
      ...
    }

In the absence of that file, autocomplete after "cob" instead looks like
this:

    Co-Authored-By: Name <user@github.com>
                    ^^^^  ^^^^ ^^^^^^^^^^
                     1     2       3

And you can tab through the "1", "2", and "3" placeholders.