fix(nvim): include handle text in handle completion autocomplete
Hadn't spotted this bug because at GitHub your handle always matches
your email address:

    @wincent     Greg Hurrell <wincent@github.com>
    ^^^^^^^^     ^^^^^^^^^^^^  ^^^^^^^^^^^^^^^^^^
    Handle       Name          Email address

As such, I could type "@wincent" and it would always match the suggested
`newText`, because "wincent" is in the email address part.

However, when the handles don't match, it would not work; ie. in the
following scenario "@wincent" wouldn't work:

    @wincent     Greg Hurrell <greg@hurrell.net>
    ^^^^^^^^     ^^^^^^^^^^^^  ^^^^^^^^^^^^^^^^
    Handle       Name          Email address

This is slightly weird because I can type something like
"@winHurrgreghurr" and it will match, but it generally yields the
expected result. (Note that including a dot in the input breaks the
autocompletion because it is not considered a keyword character; I may
have to override the keyword detection to fix that.)