fix(nvim): add "." to nvim-cmp keyword characters for handle source
So I can type a "." in something like "github.com" without interrupting
autocompletion.

For some reason, this function is _super_ fussy about the exact format
it wants things in. I tried expressing this in a bunch of ways that I
think should be equivalent, but only this one worked (and it was about
the tenth that I tried).

Breaking it down:

    [[          Lua: opening string delimiter (ignores escapes).
    \%(         Vim: start non-capturing group.
    \k          Vim: keyword character.
    \|          Vim: alternate operator.
    \.          Vim: literal dot.
    \)          Vim: end non-capturing group.
    \+          Vim: one-or-more quantifier.
    ]]          Lua: closing string delimiter.

Note that in the `gitcommit` filetype `'iskeyword'` has the default
value of:

   iskeyword=@,48-57,_,192-255

(48-57 correspond to 0-9; not sure what 192-255 correspond to.)