refactor: add local `wincent` helper and use it
This is to handle the common case of running stuff only when it's "me"
(that is, when `$USER` is "wincent"). There is an obvious combinatorial
problem with the existing helpers because they aren't composable; ie. I
can do "if arch" or "if wincent" but not "if arch and wincent" or "if
arch or wincent". So far, that's not a problem, but it might be in the
future; the main challenge is figuring out what such an API should/would
look like, as we don't want to come up with something convoluted when
simply writing some JS would suffice:

    if (
      variable('identity') === 'wincent' &&
      attributes.platform === 'arch'
    )

and so on... I may be other thinking things but I'm imagining something
like the following with implicit AND semantics:

    when('arch', 'wincent', () => {
      // do stuff...
    });

and if you wanted OR semantics, you'd represent that with an array:

    when(['arch', 'debian'], 'wincent', () => {
      // do stuff...
    });

The basic idea of the `when` helper would be to abstract over the
different kinds of checks (variable look-ups, attribute tests etc) and
just have string labels for every check that we'd care to make. So, the
above examples could effectively stand in for `if` blocks, or could be
wrangled to actually use `if` blocks like this:

    // `test` probably isn't the right word here, but something like...
    if(test(['arch', 'debian'], 'wincent')) {
      // do stuff
    }

And we'd plug either of the above into the existing `task` API like so:

    test(['arch', 'debian'], 'wincent').task('...', () => {});
    when(['arch', 'debian'], 'wincent').task('...', () => {});

Anyway, something to think about...