refactor(bitcoin): extract Bitcoin stuff into separate aspect
Given that some people use this stuff for money-related things, move it
all into a separate aspect where I can more easily cordon it off. This
stuff will only install on Arch, and only if `$USER` is "wincent". Note
the use of the default user + password, because I only want to use this
to inspect and learn about the blockchain; I have no intention of ever
using a wallet on a toy system like this one.