refactor: prepare to run some aspects in parallel
Due to the `flat()` calls, the code is currently running exactly like it
was before, but this shows that we can hack the typegen to allows us to
specify project aspects as lists:

    [
      "dotfiles",
      "node",
      "nvim"
    ]

or lists of lists (mixed, depth 1):

    [
      "dotfiles",
      ["node", "nvim"]
    ]

When I said "hack", I really meant it. I couldn't be bothered making the
JSON Schema handling totally generic, so I basically hard-coded the
desired result. It produces a type definition like this:

    export interface Project {
      'platforms': {
        'darwin'?: {
          'aspects': Array<Aspect | Array<Aspect>>;
          'variables'?: {
            [key: string]: JSONValue;
          }
        }
      ... (etc)

And assertion code like this:

    if (darwin.hasOwnProperty('aspects')) {
      const aspects: unknown = (darwin as any)['aspects'];

      assert(Array.isArray(aspects));

      for (const item of aspects) {
        if (Array.isArray(item)) {
          item.forEach(assertAspect);
        } else {
          assertAspect(item);
        }
      }
    }
    ... (etc)

In following commit, I'll see if I can refactor `main.ts` to run
sub-arrays of aspects in parallel. This is obviously going to be very
limited, such that, for example, we won't run in parallel when `--step`
is in effect (that's a rare mode of operation anyway), and we'll have to
rely on the user to not try running things in parallel that would
otherwise prompt the user (if they do show a thing like a password
prompt, I guess it will probably still work, but the output might be
confusing).