perf: run some aspects in parallel with `--parallel`
Experimental because it's not robust for reasons I haven't figured out
yet (but which I have some very solid hunches about; see below), and
that's why it is behind an off-by-default switch. You can see it running
aspects in parallel in this excerpt:

    [notice]  Task: node | make ~/n
    [notice]  Task: karabiner | make directories
    [notice]  Task: launchd | configure (global) LaunchDaemons
    [info]    Ok: file ~/n
    [notice]  Task: node | hide ~/n
    [info]    Ok: file ~/.config
    [notice]  Changed: command `chflags hidden ~/n`
    [notice]  Task: node | install Node.js v18.0.0
    [info]    Ok: file ~/.config/karabiner
    [info]    Skipped: command `/Users/wincent/code/wincent/vendor/n/bin/n 18.0.0`
    [notice]  Task: node | install global packages
    [notice]  Summary: changed=1 failed=0 ok=12 skipped=4 elapsed=1.69s

But you can also see a `stat` call mysteriously returning an empty
stdout string, via this logging I added:

    stdout: "" for args ["-f","%Lp%n%HT%n%Su%n%Sg%n%Y","/Users/wincent/bin"]

That in turn causes this error:

    [error]   TypeError: Cannot read properties of undefined (reading 'toLowerCase')

because we try to split that string and extract the type information
("directory") out of it. The output should be:

    755
    Directory
    wincent
    staff

I'll keep digging, but it's possible this feature needs to forever
remain dormant behind a flag. I know of some definitely not-safe
singleton state (eg. in the Context class and probably other places too)
that will need to be refactored in order for all of this to work
solidly, and I may want to implement a mutex around very critical
sections too (specifically, thinking about anything that prints a
prompt; there should only be one of those at a time).